created resolver directory to modularize
