+++ src/models/index.js
@@ -1,6 +1,8 @@
+/*import Note*/
 const Note = require('../note');
+/*store Note in models*/
 const models = {
   Note
 };
-
+/*export models*/
 module.exports = models;
