+++ src/resolvers/query.js
@@ -1,0 +1,9 @@
+/*export query resolvers*/
+module.exports = {
+  notes: async (parent, args, {models}) => {
+    return models.Note.find();
+  },
+  note: async (parent, args,{models}) => {
+    return models.Note.findById(args.id);
+  }
+};
